backend/requirements.txt
# FastMCP and core dependencies
fastmcp==0.1.0
mcp==0.1.0
pydantic==2.5.0
uvicorn==0.24.0

# AI and ML dependencies
google-generativeai==0.3.2
openai==1.3.0
mem0ai==0.1.10
chromadb==0.4.18
sentence-transformers==2.2.2

# Audio processing
speechrecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
pygame==2.5.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Database and vector search
sqlite3-fts4==1.0.3
faiss-cpu==1.7.4
sqlalchemy==2.0.23

# Web and API dependencies
httpx==0.25.2
fastapi==0.104.1
flask==3.0.3
flask-cors==5.0.0
requests==2.31.0
aiofiles==23.2.1
gunicorn==22.0.0

# Weather and location
pyowm==3.3.0
geopy==2.4.0

# Health and medical
pandas==2.1.4
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
asyncio-throttle==1.0.2
pathlib2==2.3.7
